Update adapter package names.
