+++ components/org.wso2.identity.outbound.adapter.common/pom.xml
@@ -22,7 +22,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>identity-outbound-adapters</artifactId>
-        <groupId>com.wso2.identity.outbound.adapters</groupId>
+        <groupId>org.wso2.identity.outbound.adapters</groupId>
         <version>1.0.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
