+++ pom.xml
@@ -24,19 +24,19 @@
         <version>1.4</version>
     </parent>
 
-    <groupId>com.wso2.identity.outbound.adapters</groupId>
+    <groupId>org.wso2.identity.outbound.adapters</groupId>
     <modelVersion>4.0.0</modelVersion>
     <version>1.0.0-SNAPSHOT</version>
     <artifactId>identity-outbound-adapters</artifactId>
 
     <packaging>pom</packaging>
     <name>Outbound Adapters</name>
-    <description>Maintain output event adapters which do not depend on Siddhi event stream techniques.</description>
+    <description>Maintain outbound adapters to be integrate with different providers</description>
     <url>http://wso2.org</url>
 
     <modules>
         <module>components/org.wso2.identity.outbound.adapter.common</module>
-        <module>components/org.wso2.carbon.event.outbound.adapter.websubhub</module>
+        <module>components/org.wso2.identity.outbound.adapter.websubhub</module>
     </modules>
 
     <scm>
@@ -80,7 +80,7 @@
                 </exclusions>
             </dependency>
             <dependency>
-                <groupId>com.wso2.identity.outbound.adapters</groupId>
+                <groupId>org.wso2.identity.outbound.adapters</groupId>
                 <artifactId>org.wso2.identity.outbound.adapter.common</artifactId>
                 <version>${project.version}</version>
             </dependency>
